+++ src/app/features/pokemon/presentation/components/poke-grid.tsx
@@ -1,10 +1,10 @@
 import { Suspense, useEffect, useRef } from 'react'
 import { useVirtualizer, Virtualizer } from '@tanstack/react-virtual'
 import type { PokemonListItem } from '@/app/features/pokemon/domain/models/pokemon-list-models'
-import { PokeCard } from './PokeCard'
-import { SkeletonPokeCard } from './SkeletonPokeCard'
+import { PokeCard } from './poke-card'
+import { SkeletonPokeCard } from './skeleton-poke-card'
 
-import { useMediaQuery } from '@/app/shared/hooks/useMidiaQuery'
+import { useMediaQuery } from '@/app/shared/hooks/use-midia-query'
 
 type PokemonGridProps = {
   results: PokemonListItem[] | undefined
