+++ src/app/features/pokemon/presentation/components/pokemon-not-found.tsx
@@ -1,4 +1,4 @@
-import { useTranslation } from '@/app/shared/hooks/useTranslation'
+import { useTranslation } from '@/app/shared/hooks/use-translation'
 
 export const PokemonNotFound = () => {
   const { t } = useTranslation('common')
