+++ src/app/features/pokemon/presentation/components/skeleton-poke-list.tsx
@@ -1,4 +1,4 @@
-import { SkeletonPokeCard } from './SkeletonPokeCard'
+import { SkeletonPokeCard } from './skeleton-poke-card'
 
 export const SkeletonPokeList = () => {
   return (
