+++ src/app/features/pokemon/presentation/screens/pokemon-search-screen.tsx
@@ -1,5 +1,5 @@
 import { Link, useParams } from '@tanstack/react-router'
-import { PokeCard } from '../components/PokeCard'
+import { PokeCard } from '../components/poke-card'
 import { ArrowIcon } from '@/app/shared/components/icons/arrow'
 import { Button } from '@/app/shared/components/ui/button'
 import { useTranslation } from 'react-i18next'
